+++ components/FileFormats.vue
@@ -1,6 +1,6 @@
 <template>
 	<div class="vue-component file-formats">
-		<SearchableList :data="fileFormats" summaryKey="title" :showSummaryOnExpand="false" :externalSearchTerm="searchTerm" :sort="sort" :offerDetails="offerDetails" :heading="heading" :collapsed="collapsed">
+		<SearchableList :data="fileFormats" summaryKey="title" :showSummaryOnExpand="false" :externalSearchTerm="searchTerm" :sort="sort" :offerDetails="offerDetails" :heading="heading" :collapsed="collapsed" @detailsToggled="detailsToggled">
 			<template #heading="scope"><slot name="heading" v-bind="scope" /></template>
 			<template #summary="slot">
 				<slot name="summary" v-bind="slot">
@@ -97,6 +97,11 @@ export default {
 	},
 	beforeCreate() {
 		Utils.enableHtmlProps(this);
+	},
+	methods: {
+		detailsToggled(...args) {
+			this.$emit('detailsToggled', ...args);
+		}
 	}
 }
 </script>